pyproject.toml
[project]
name = "options-chatbot"
version = "0.1.0"
description = "Options strategy explainer and payoff calculator with live premiums"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "cachetools>=5.3",
  "google-cloud-aiplatform>=1.67",
]

[tool.uv]
package = false
